third_party/WebKit/LayoutTests/http/tests/inspector/sources/debugger/pause-in-removed-frame.html

<html>
<head>
<script src="../../inspector-test.js"></script>
<script src="../../debugger-test.js"></script>
<script>

window.removeIframe = function()
{
  var child = document.getElementById('child');
  child.parentNode.removeChild(child);
  debugger;
};

function testFunction()
{
    var childWindow = document.getElementById("child").contentWindow;
    childWindow.sendRequest();
}

var test = function()
{
    InspectorTest.startDebuggerTest(step1);

    function step1()
    {
        InspectorTest.runTestFunctionAndWaitUntilPaused(step2);
    }

    function step2(callFrames)
    {
        InspectorTest.captureStackTrace(callFrames);
        InspectorTest.completeDebuggerTest();
    }
}

</script>
</head>

<body>
<div>Parent Frame</div>
<iframe id="child" src="./resources/child.html" onload="runTest()"></iframe>
<p>
Tests "pause" functionality in detached frame.
</p>

</body>
</html>
