third_party/WebKit/LayoutTests/http/tests/inspector/sources/debugger/pause-in-removed-frame-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check http://xhr.spec.whatwg.org/.
CONSOLE ERROR: line 16: Uncaught NetworkError: Failed to execute 'send' on 'XMLHttpRequest': Failed to load 'http://127.0.0.1:8000/inspector/sources/debugger/resources/resource/unreachable'.
Parent Frame
Tests "pause" functionality in detached frame.

Set timer for test function.
Script execution paused.
Call stack:
    0) window.removeIframe (pause-in-removed-frame.html:11)
    1) onXhrLoad (child.html:12)
    2) sendRequest (child.html:16)
    3) testFunction (pause-in-removed-frame.html:17)
Script execution resumed.
